+++ crates/testing-utils/src/helpers.rs
@@ -20,16 +20,16 @@ use crate::{
     ChainSpecType,
 };
 use entropy_protocol::PartyId;
-use subxt::{backend::legacy::LegacyRpcMethods, OnlineClient};
+use rand::{rngs::StdRng, SeedableRng};
+use subxt::{backend::legacy::LegacyRpcMethods, utils::AccountId32, OnlineClient};
+pub use tdx_quote::encode_verifying_key;
 
 /// A helper for setting up tests which starts both a set of TS servers and a chain node and returns
 /// the chain API as well as IP addresses and PartyId of the started validators
 pub async fn spawn_tss_nodes_and_start_chain(
-    add_parent_key: bool,
     chain_spec_type: ChainSpecType,
 ) -> (OnlineClient<EntropyConfig>, LegacyRpcMethods<EntropyConfig>, Vec<String>, Vec<PartyId>) {
-    let (validator_ips, validator_ids) =
-        spawn_testing_validators(add_parent_key, chain_spec_type.clone()).await;
+    let (validator_ips, validator_ids) = spawn_testing_validators(chain_spec_type).await;
 
     let (api, rpc) = match chain_spec_type {
         ChainSpecType::Development => {
@@ -51,4 +51,19 @@ pub async fn spawn_tss_nodes_and_start_chain(
         },
     };
     (api, rpc, validator_ips, validator_ids)
+}
+
+/// Get the mock PCK that will be used for a given TSS account ID
+pub fn derive_mock_pck_verifying_key(tss_account_id: &AccountId32) -> tdx_quote::VerifyingKey {
+    let mut pck_seeder = StdRng::from_seed(tss_account_id.0);
+    let pck = tdx_quote::SigningKey::random(&mut pck_seeder);
+    tdx_quote::VerifyingKey::from(pck)
+}
+
+/// For each test TSS account, display the encoded mock PCK
+pub fn print_test_pck_verifying_keys() {
+    for tss_account in crate::constants::TSS_ACCOUNTS.iter() {
+        let pck = derive_mock_pck_verifying_key(tss_account);
+        println!("{:?}", encode_verifying_key(&pck));
+    }
 }